Removes usage of C preprocessor, adds CMake build
I removed the use of the preprocessor by moving the number of
derivatives to a separate module file. This module file as-is is
incomplete, and needs to be run through CMake's configure_file()
command. The CMake build contains the DNAD_NDERIV cache variable, which
must be set at build time. If it is not set, the build will error out.